refactor: extract GitHubExportProvider+Duplicates (#639 slice C)
Third slice of #639 (after PR #875 WireTypes and PR #879 Networking).
Moves the duplicate-detection + reconcile-pending cluster into a
dedicated Sources/BugNarrator/Services/GitHubExportProvider+Duplicates.swift
extension file.

Moved (6 members):
- func findOpenIssues(matching:configuration:) — already public, unchanged
- makeSearchRequest — stays private (single-caller findOpenIssues moves with)
- makeExportFingerprintSearchRequest — stays private (single-caller findExportedIssue moves with)
- existingExportResult — widen private -> internal (called from export() in base file)
- reconcilePendingExport — widen private -> internal (called from createWithRetry in base file, IC/DUP boundary)
- findExportedIssue — stays private (single-caller reconcilePendingExport moves with)

Additional visibility widening required (found by build-time check;
codex pre-review missed the actor stored-property dependencies):
- session (URLSession): private -> internal (used by findExportedIssue)
- receiptStore (ExportReceiptStoring): private -> internal (used by reconcilePendingExport)

The 2 stored-property widenings were not in the AC and were discovered
by the compiler. Minimal necessary widening: retryConfiguration, logger,
and annotationRenderer stay private.

Line-count delta:
- GitHubExportProvider.swift: 665 -> 519 (-146 lines)
- GitHubExportProvider+Duplicates.swift (new): 155 lines
- xcodeproj: +4 (xcodegen)

Test coverage: GitHubExportProviderTests continues to pass unchanged.

Closes #880.
